Update Mywork.py
The first two questions worked out smoothly for me. I was able to use loops to find the numbers that I need based on the pattern of the numbers. I spent a lot of time on Question 3, because I do not know how to find prime numbers. There are no visible patterns to find prime numbers except that they can only be divided by 1 and themselves. It is even harder to find the largest prime factor. I finally came up with an idea to use loops and if else statement to test if n can be divided by prime numbers smaller than the square root of n. I am still not a hundred percent sure if question three works. I would like to discuss it with my peers in class.