Work Session2 reflection
For this work session, I only solved one problem(problem4)  because it is very complicated. Initially I was stuck because I did not know how to find a palindrome, not to mention the biggest palindrome number through the multiplication of three digit numbers. Then after some brainstorming, I first wrote the helper function to determine a palindrome using the property that the palindrome reads the same both ways. I use int(length/2) to find the middle index of each number and check it. Then I came up with the idea to loop through all three digit numbers, multiply them and check if their product is a palindrome. I call on my helper function in the loop to check the products. This loop requires a lot of calculations and my python software got stuck while running the code. I have to quit and restart.