Work Session3 reflection
In this work session, I solved Question 5 of project Euler. First I factored all numbers from 1 to 20 on paper, and found that the 11 to 20 contains all other numbers as their factors. For instance 15 can be factored into 3 and 5. Therefore, I decide to only consider numbers 11 to 20 in my code to reduce the amount of calculations. Similar to Question4 last time, I first completed the helper function to check if the number can be evenly divided by numbers from 11 to 20. Then  I multiplied 11 to 20 to find the maximum range of the main loop. I used the for loop to check if all numbers from 0 to the product can be evenly divided using the helper function. If the helper function returns True, the number is added into a new list. Finally, the minimum number in the list is returned. In this session, I mainly struggled with writing the various loops and I was stuck several times.